Naive refactoring just to break the big main.cc into more manageable, smaller files
