+++ .gitignore
@@ -40,3 +40,4 @@ CMakeCache.txt
 
 CMakeFiles
 .cmake
+venv
